% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Example data for converting mislabeled female bodymaps to standard}
\format{A data frame with 3 rows and 2 variables:
\describe{
  \item{GENDER}{gender of the patient}
  \item{BODYMAP_CSV}{bodymap endorsements in comma separated form}
}}
\usage{
exampledata
}
\description{
A dataset containing gender and bodymap segment endorsements for multiple patients
}
\keyword{datasets}
