% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_bodymap.R
\name{prep_bodymaps}
\alias{prep_bodymaps}
\title{prep_bodymaps
converts a single charcter vector of bodymaps into a list of character vectors, each a bodymap}
\usage{
prep_bodymaps(maps)
}
\arguments{
\item{maps}{a character vector containing the endorsed bodymap segments of patients in csv form}
}
\value{
a list of character vectors, where each vector contains the patient's endorsed segments
}
\description{
prep_bodymaps
converts a single charcter vector of bodymaps into a list of character vectors, each a bodymap
}
\examples{
prep_bodymaps(as.character(exampledata$BODYMAP_CSV))
}
